mv store code into Store namespace and move model code up to Wallet namespace
