+++ nix/materialized/stack-nix/cardano-wallet-core.nix
@@ -167,6 +167,7 @@
           "Cardano/Pool/Rank"
           "Cardano/Pool/Rank/Likelihood"
           "Cardano/Wallet"
+          "Cardano/Wallet/Address/Book"
           "Cardano/Wallet/Address/Pool"
           "Cardano/Wallet/Api"
           "Cardano/Wallet/Api/Aeson/Variant"
@@ -176,6 +177,7 @@
           "Cardano/Wallet/Api/Server/Tls"
           "Cardano/Wallet/Api/Types"
           "Cardano/Wallet/Api/Types/SchemaMetadata"
+          "Cardano/Wallet/Checkpoints"
           "Cardano/Wallet/CoinSelection"
           "Cardano/Wallet/CoinSelection/Internal"
           "Cardano/Wallet/CoinSelection/Internal/Balance"
@@ -183,16 +185,14 @@
           "Cardano/Wallet/CoinSelection/Internal/Context"
           "Cardano/Wallet/Compat"
           "Cardano/Wallet/DB"
-          "Cardano/Wallet/DB/Checkpoints"
           "Cardano/Wallet/DB/Pure/Implementation"
           "Cardano/Wallet/DB/Pure/Layer"
-          "Cardano/Wallet/DB/Sqlite/AddressBook"
-          "Cardano/Wallet/DB/Sqlite/CheckpointsOld"
           "Cardano/Wallet/DB/Layer"
           "Cardano/Wallet/DB/Sqlite/Migration"
           "Cardano/Wallet/DB/Sqlite/Schema"
           "Cardano/Wallet/DB/Sqlite/Stores"
           "Cardano/Wallet/DB/Sqlite/Types"
+          "Cardano/Wallet/DB/Store/Checkpoints"
           "Cardano/Wallet/DB/WalletState"
           "Cardano/Wallet/Logging"
           "Cardano/Wallet/Network"
